#47 change int to float

+++ RecruitmentSystemAPI/Repositories/LabourerJobsRepo.cs
@@ -56,20 +56,13 @@ namespace RecruitmentSystemAPI.Repositories
             return labourerJobVM;
         }
 
-        public void UpdateJobRating(int idToGrade, int rating, string usesrId)
+        public void UpdateJobRating(int idToGrade, float rating, string usesrId)
         {
             var x = idToGrade;
             var jobToRate = _context.LabourerJobs.Where(ls => ls.Id == idToGrade && ls.Labourer.UserId == usesrId).FirstOrDefault();
-            //var jobToRate = _context.LabourerJobs.FirstOrDefault(l => l.JobId == 1);
-            //jobToRate.Id = jobToRate.Id;
-            //jobToRate.JobId = jobToRate.JobId;
             jobToRate.JobRating = rating;
-            //jobToRate.QualityRating = 
             _context.Update(jobToRate);
             _context.SaveChanges();
-
         }
-
-
     }
 }
